completed the saving for the player and the car
