--- Assets/PlayerPostion.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f4b1dd31f01e95d48afc869f163ac4f7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
