+++ Assets/Scripts/AdvancedSaving/SavePlayerPos.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b64148e41bab875439768f6c0eb8c648
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
